+++ engine/sc_mmo_champion.cpp
@@ -265,7 +265,7 @@ static bool parse_item_stats( item_t&     item,
   std::vector<std::string> descriptions;
   get_tti_value( descriptions, node, "tti-stat"  );
   get_tti_value( descriptions, node, "tti-stats" );
-  int num_descriptions = descriptions.size();
+  size_t num_descriptions = descriptions.size();
   for ( int i=0; i < num_descriptions; i++ )
   {
     armory_t::fuzzy_stats( item.armory_stats_str, descriptions[ i ] );
