+++ engine/sc_option.cpp
@@ -277,7 +277,7 @@ bool option_t::parse_file( sim_t* sim,
     if ( first )
     {
       // Skip the Windows UTF-8 magic cookie.
-      uint32_t len = strlen( b );
+      size_t len = strlen( b );
       if ( ( len >= 3 ) && ( ( unsigned char ) b[ 0 ] == 0xEF ) && ( ( unsigned char ) b[ 1 ] == 0xBB ) && ( ( unsigned char ) b[ 2 ] == 0xBF ) )
       {
         b += 3;
